Simplify the UCD codegen Python script
