+++ src/core/unicode/codegen.py
@@ -4,298 +4,127 @@ import re
 import sys
 
 LINE = re.compile(r"^([0-9A-Fa-f]+)(?:\.\.([0-9A-Fa-f]+))?\s*;\s*(\S+)")
-MISSING = re.compile(
-    r"^#\s*@missing:\s*([0-9A-Fa-f]+)\.\.([0-9A-Fa-f]+)\s*;\s*(\S+)"
-)
+MISSING_PREFIX = re.compile(r"^#\s*@missing:\s*")
 
 TOTAL_CODEPOINTS = 0x110000
 PAGE_SHIFT = 10
 PAGE_SIZE = 1 << PAGE_SHIFT
 NUM_PAGES = TOTAL_CODEPOINTS // PAGE_SIZE
 
-JOINING_TYPE_VALUES = {
-    "U": 0,
-    "T": 1,
-    "L": 2,
-    "R": 3,
-    "D": 4,
-    "C": 5,
-}
+# Per-property canonical order. Position in this list defines the integer
+# value of the matching C++ enum entry. PropertyValueAliases.txt supplies
+# the short/long alias mappings at codegen time, so we only need to
+# declare one form per value here.
 
-JOINING_TYPE_FULL_NAMES = {
-    "Non_Joining": 0,
-    "Transparent": 1,
-    "Left_Joining": 2,
-    "Right_Joining": 3,
-    "Dual_Joining": 4,
-    "Join_Causing": 5,
-}
+JOINING_TYPE_ORDER = ["U", "T", "L", "R", "D", "C"]
 
-BIDI_CLASS_VALUES = {
-    "L": 0,
-    "R": 1,
-    "AL": 2,
-    "EN": 3,
-    "ES": 4,
-    "ET": 5,
-    "AN": 6,
-    "CS": 7,
-    "NSM": 8,
-    "BN": 9,
-    "B": 10,
-    "S": 11,
-    "WS": 12,
-    "ON": 13,
-    "LRE": 14,
-    "LRO": 15,
-    "RLE": 16,
-    "RLO": 17,
-    "PDF": 18,
-    "LRI": 19,
-    "RLI": 20,
-    "FSI": 21,
-    "PDI": 22,
-}
+BIDI_CLASS_ORDER = [
+    "L", "R", "AL", "EN", "ES", "ET", "AN", "CS", "NSM", "BN",
+    "B", "S", "WS", "ON", "LRE", "LRO", "RLE", "RLO", "PDF",
+    "LRI", "RLI", "FSI", "PDI",
+]
 
-BIDI_CLASS_FULL_NAMES = {
-    "Left_To_Right": 0,
-    "Right_To_Left": 1,
-    "Arabic_Letter": 2,
-    "European_Number": 3,
-    "European_Separator": 4,
-    "European_Terminator": 5,
-    "Arabic_Number": 6,
-    "Common_Separator": 7,
-    "Nonspacing_Mark": 8,
-    "Boundary_Neutral": 9,
-    "Paragraph_Separator": 10,
-    "Segment_Separator": 11,
-    "White_Space": 12,
-    "Other_Neutral": 13,
-    "Left_To_Right_Embedding": 14,
-    "Left_To_Right_Override": 15,
-    "Right_To_Left_Embedding": 16,
-    "Right_To_Left_Override": 17,
-    "Pop_Directional_Format": 18,
-    "Left_To_Right_Isolate": 19,
-    "Right_To_Left_Isolate": 20,
-    "First_Strong_Isolate": 21,
-    "Pop_Directional_Isolate": 22,
-}
-
-UNICODE_SCRIPT_VALUES = {
-    "Adlam": 0,
-    "Ahom": 1,
-    "Anatolian_Hieroglyphs": 2,
-    "Arabic": 3,
-    "Armenian": 4,
-    "Avestan": 5,
-    "Balinese": 6,
-    "Bamum": 7,
-    "Bassa_Vah": 8,
-    "Batak": 9,
-    "Bengali": 10,
-    "Beria_Erfe": 11,
-    "Bhaiksuki": 12,
-    "Bopomofo": 13,
-    "Brahmi": 14,
-    "Braille": 15,
-    "Buginese": 16,
-    "Buhid": 17,
-    "Canadian_Aboriginal": 18,
-    "Carian": 19,
-    "Caucasian_Albanian": 20,
-    "Chakma": 21,
-    "Cham": 22,
-    "Cherokee": 23,
-    "Chorasmian": 24,
-    "Common": 25,
-    "Coptic": 26,
-    "Cuneiform": 27,
-    "Cypriot": 28,
-    "Cypro_Minoan": 29,
-    "Cyrillic": 30,
-    "Deseret": 31,
-    "Devanagari": 32,
-    "Dives_Akuru": 33,
-    "Dogra": 34,
-    "Duployan": 35,
-    "Egyptian_Hieroglyphs": 36,
-    "Elbasan": 37,
-    "Elymaic": 38,
-    "Ethiopic": 39,
-    "Garay": 40,
-    "Georgian": 41,
-    "Glagolitic": 42,
-    "Gothic": 43,
-    "Grantha": 44,
-    "Greek": 45,
-    "Gujarati": 46,
-    "Gunjala_Gondi": 47,
-    "Gurmukhi": 48,
-    "Gurung_Khema": 49,
-    "Han": 50,
-    "Hangul": 51,
-    "Hanifi_Rohingya": 52,
-    "Hanunoo": 53,
-    "Hatran": 54,
-    "Hebrew": 55,
-    "Hiragana": 56,
-    "Imperial_Aramaic": 57,
-    "Inherited": 58,
-    "Inscriptional_Pahlavi": 59,
-    "Inscriptional_Parthian": 60,
-    "Javanese": 61,
-    "Kaithi": 62,
-    "Kannada": 63,
-    "Katakana": 64,
-    "Kawi": 65,
-    "Kayah_Li": 66,
-    "Kharoshthi": 67,
-    "Khitan_Small_Script": 68,
-    "Khmer": 69,
-    "Khojki": 70,
-    "Khudawadi": 71,
-    "Kirat_Rai": 72,
-    "Lao": 73,
-    "Latin": 74,
-    "Lepcha": 75,
-    "Limbu": 76,
-    "Linear_A": 77,
-    "Linear_B": 78,
-    "Lisu": 79,
-    "Lycian": 80,
-    "Lydian": 81,
-    "Mahajani": 82,
-    "Makasar": 83,
-    "Malayalam": 84,
-    "Mandaic": 85,
-    "Manichaean": 86,
-    "Marchen": 87,
-    "Masaram_Gondi": 88,
-    "Medefaidrin": 89,
-    "Meetei_Mayek": 90,
-    "Mende_Kikakui": 91,
-    "Meroitic_Cursive": 92,
-    "Meroitic_Hieroglyphs": 93,
-    "Miao": 94,
-    "Modi": 95,
-    "Mongolian": 96,
-    "Mro": 97,
-    "Multani": 98,
-    "Myanmar": 99,
-    "Nabataean": 100,
-    "Nag_Mundari": 101,
-    "Nandinagari": 102,
-    "New_Tai_Lue": 103,
-    "Newa": 104,
-    "Nko": 105,
-    "Nushu": 106,
-    "Nyiakeng_Puachue_Hmong": 107,
-    "Ogham": 108,
-    "Ol_Chiki": 109,
-    "Ol_Onal": 110,
-    "Old_Hungarian": 111,
-    "Old_Italic": 112,
-    "Old_North_Arabian": 113,
-    "Old_Permic": 114,
-    "Old_Persian": 115,
-    "Old_Sogdian": 116,
-    "Old_South_Arabian": 117,
-    "Old_Turkic": 118,
-    "Old_Uyghur": 119,
-    "Oriya": 120,
-    "Osage": 121,
-    "Osmanya": 122,
-    "Pahawh_Hmong": 123,
-    "Palmyrene": 124,
-    "Pau_Cin_Hau": 125,
-    "Phags_Pa": 126,
-    "Phoenician": 127,
-    "Psalter_Pahlavi": 128,
-    "Rejang": 129,
-    "Runic": 130,
-    "Samaritan": 131,
-    "Saurashtra": 132,
-    "Sharada": 133,
-    "Shavian": 134,
-    "Siddham": 135,
-    "Sidetic": 136,
-    "SignWriting": 137,
-    "Sinhala": 138,
-    "Sogdian": 139,
-    "Sora_Sompeng": 140,
-    "Soyombo": 141,
-    "Sundanese": 142,
-    "Sunuwar": 143,
-    "Syloti_Nagri": 144,
-    "Syriac": 145,
-    "Tagalog": 146,
-    "Tagbanwa": 147,
-    "Tai_Le": 148,
-    "Tai_Tham": 149,
-    "Tai_Viet": 150,
-    "Tai_Yo": 151,
-    "Takri": 152,
-    "Tamil": 153,
-    "Tangsa": 154,
-    "Tangut": 155,
-    "Telugu": 156,
-    "Thaana": 157,
-    "Thai": 158,
-    "Tibetan": 159,
-    "Tifinagh": 160,
-    "Tirhuta": 161,
-    "Todhri": 162,
-    "Tolong_Siki": 163,
-    "Toto": 164,
-    "Tulu_Tigalari": 165,
-    "Ugaritic": 166,
-    "Unknown": 167,
-    "Vai": 168,
-    "Vithkuqi": 169,
-    "Wancho": 170,
-    "Warang_Citi": 171,
-    "Yezidi": 172,
-    "Yi": 173,
-    "Zanabazar_Square": 174,
-}
+UNICODE_SCRIPT_ORDER = [
+    "Adlam", "Ahom", "Anatolian_Hieroglyphs", "Arabic", "Armenian",
+    "Avestan", "Balinese", "Bamum", "Bassa_Vah", "Batak", "Bengali",
+    "Beria_Erfe", "Bhaiksuki", "Bopomofo", "Brahmi", "Braille",
+    "Buginese", "Buhid", "Canadian_Aboriginal", "Carian",
+    "Caucasian_Albanian", "Chakma", "Cham", "Cherokee", "Chorasmian",
+    "Common", "Coptic", "Cuneiform", "Cypriot", "Cypro_Minoan",
+    "Cyrillic", "Deseret", "Devanagari", "Dives_Akuru", "Dogra",
+    "Duployan", "Egyptian_Hieroglyphs", "Elbasan", "Elymaic",
+    "Ethiopic", "Garay", "Georgian", "Glagolitic", "Gothic", "Grantha",
+    "Greek", "Gujarati", "Gunjala_Gondi", "Gurmukhi", "Gurung_Khema",
+    "Han", "Hangul", "Hanifi_Rohingya", "Hanunoo", "Hatran", "Hebrew",
+    "Hiragana", "Imperial_Aramaic", "Inherited", "Inscriptional_Pahlavi",
+    "Inscriptional_Parthian", "Javanese", "Kaithi", "Kannada", "Katakana",
+    "Kawi", "Kayah_Li", "Kharoshthi", "Khitan_Small_Script", "Khmer",
+    "Khojki", "Khudawadi", "Kirat_Rai", "Lao", "Latin", "Lepcha", "Limbu",
+    "Linear_A", "Linear_B", "Lisu", "Lycian", "Lydian", "Mahajani",
+    "Makasar", "Malayalam", "Mandaic", "Manichaean", "Marchen",
+    "Masaram_Gondi", "Medefaidrin", "Meetei_Mayek", "Mende_Kikakui",
+    "Meroitic_Cursive", "Meroitic_Hieroglyphs", "Miao", "Modi",
+    "Mongolian", "Mro", "Multani", "Myanmar", "Nabataean", "Nag_Mundari",
+    "Nandinagari", "New_Tai_Lue", "Newa", "Nko", "Nushu",
+    "Nyiakeng_Puachue_Hmong", "Ogham", "Ol_Chiki", "Ol_Onal",
+    "Old_Hungarian", "Old_Italic", "Old_North_Arabian", "Old_Permic",
+    "Old_Persian", "Old_Sogdian", "Old_South_Arabian", "Old_Turkic",
+    "Old_Uyghur", "Oriya", "Osage", "Osmanya", "Pahawh_Hmong",
+    "Palmyrene", "Pau_Cin_Hau", "Phags_Pa", "Phoenician",
+    "Psalter_Pahlavi", "Rejang", "Runic", "Samaritan", "Saurashtra",
+    "Sharada", "Shavian", "Siddham", "Sidetic", "SignWriting", "Sinhala",
+    "Sogdian", "Sora_Sompeng", "Soyombo", "Sundanese", "Sunuwar",
+    "Syloti_Nagri", "Syriac", "Tagalog", "Tagbanwa", "Tai_Le", "Tai_Tham",
+    "Tai_Viet", "Tai_Yo", "Takri", "Tamil", "Tangsa", "Tangut", "Telugu",
+    "Thaana", "Thai", "Tibetan", "Tifinagh", "Tirhuta", "Todhri",
+    "Tolong_Siki", "Toto", "Tulu_Tigalari", "Ugaritic", "Unknown", "Vai",
+    "Vithkuqi", "Wancho", "Warang_Citi", "Yezidi", "Yi", "Zanabazar_Square",
+]
 
 
-def parse_value(path, line_number, raw_value, value_parser, label):
-    try:
-        return value_parser(raw_value)
-    except (KeyError, ValueError) as error:
-        raise ValueError(
-            f"{path}:{line_number}: invalid {label} value "
-            f"{raw_value!r}: {error}"
-        ) from error
-
-
-def parse_missing(path, name_parser):
-    missing = []
-    with open(path) as source:
-        for line_number, line in enumerate(source, start=1):
-            stripped = line.strip()
-            match = MISSING.match(stripped)
-            if not match:
+def parse_alias_lines(aliases_path, property_short):
+    rows = []
+    with open(aliases_path) as source:
+        for line in source:
+            stripped = line.split("#", 1)[0].strip()
+            if not stripped:
                 continue
-            first = int(match.group(1), 16)
-            last = int(match.group(2), 16)
-            value = parse_value(
-                path, line_number, match.group(3), name_parser, "@missing"
+            parts = [part.strip() for part in stripped.split(";")]
+            if parts[0] == property_short:
+                rows.append([field for field in parts[1:] if field])
+    return rows
+
+
+def build_value_map(aliases_path, property_short, canonical_order=None, ignore=()):
+    """Build {form: int} for a property. With canonical_order, each row's
+    integer is its canonical's position in that list; without, the row's
+    first field is read as the integer directly (used for ccc)."""
+    canonical_to_int = (
+        {name: index for index, name in enumerate(canonical_order)}
+        if canonical_order is not None
+        else None
+    )
+    ignored = set(ignore)
+    result = {}
+    unmatched = []
+    for row in parse_alias_lines(aliases_path, property_short):
+        if any(field in ignored for field in row):
+            continue
+        if canonical_to_int is None:
+            value = int(row[0])
+        else:
+            value = next(
+                (canonical_to_int[field] for field in row if field in canonical_to_int),
+                None,
             )
-            missing.append((first, last, value))
-    return missing
+            if value is None:
+                unmatched.append(row)
+                continue
+        for field in row:
+            result[field] = value
+    if unmatched:
+        raise ValueError(
+            f"{aliases_path}: property {property_short!r} has values not "
+            f"declared in canonical order: {unmatched}"
+        )
+    return result
 
 
-def parse_ranges(path, value_parser):
-    ranges = []
+def parse_file(path, value_map):
+    """Read a UCD file and return a list of (first, last, value) entries
+    in file order. UAX #44 puts @missing lines before data lines, so the
+    file-order list can be applied directly to the dense values array."""
+    entries = []
     with open(path) as source:
         for line_number, line in enumerate(source, start=1):
             stripped = line.strip()
-            if not stripped or stripped.startswith("#"):
+            if not stripped:
                 continue
+            if stripped.startswith("#"):
+                prefix = MISSING_PREFIX.match(stripped)
+                if not prefix:
+                    continue
+                stripped = stripped[prefix.end():]
             match = LINE.match(stripped)
             if not match:
                 raise ValueError(
@@ -303,25 +132,21 @@ def parse_ranges(path, value_parser):
                 )
             first = int(match.group(1), 16)
             last = int(match.group(2), 16) if match.group(2) else first
-            value = parse_value(
-                path, line_number, match.group(3), value_parser, "data"
-            )
-            ranges.append((first, last, value))
-    return ranges
+            raw_value = match.group(3)
+            try:
+                value = value_map[raw_value]
+            except KeyError as error:
+                raise ValueError(
+                    f"{path}:{line_number}: invalid value {raw_value!r}: {error}"
+                ) from error
+            entries.append((first, last, value))
+    return entries
 
 
-def build_dense_values(missing_ranges, ranges):
+def build_pages(entries):
     values = [0] * TOTAL_CODEPOINTS
-    for first, last, value in missing_ranges:
-        for codepoint in range(first, last + 1):
-            values[codepoint] = value
-    for first, last, value in ranges:
-        for codepoint in range(first, last + 1):
-            values[codepoint] = value
-    return values
-
-
-def build_pages(values):
+    for first, last, value in entries:
+        values[first : last + 1] = [value] * (last - first + 1)
     page_to_id = {}
     unique_pages = []
     stage1 = []
@@ -347,7 +172,6 @@ def emit_property(output, prefix, stage1, unique_pages):
     )
     emit_row(output, stage1)
     output.write("};\n\n")
-
     stage2_size = len(unique_pages) * PAGE_SIZE
     output.write(
         f"constexpr std::uint8_t {prefix}_STAGE2[{stage2_size}] = {{\n"
@@ -357,18 +181,12 @@ def emit_property(output, prefix, stage1, unique_pages):
     output.write("};\n\n")
 
 
-def generate_property(input_path, data_parser, missing_parser):
-    missing = parse_missing(input_path, missing_parser)
-    ranges = parse_ranges(input_path, data_parser)
-    values = build_dense_values(missing, ranges)
-    return build_pages(values)
-
-
 def main():
-    if len(sys.argv) != 6:
+    if len(sys.argv) != 7:
         print(
             f"Usage: {sys.argv[0]} "
             "<output.h> "
+            "<PropertyValueAliases.txt> "
             "<DerivedCombiningClass.txt> "
             "<DerivedJoiningType.txt> "
             "<DerivedBidiClass.txt> "
@@ -378,47 +196,26 @@ def main():
         sys.exit(1)
 
     output_path = sys.argv[1]
-    combining_class_input = sys.argv[2]
-    joining_type_input = sys.argv[3]
-    bidi_class_input = sys.argv[4]
-    script_input = sys.argv[5]
+    aliases_path = sys.argv[2]
 
-    combining_class_stage1, combining_class_pages = generate_property(
-        combining_class_input,
-        data_parser=int,
-        missing_parser=lambda raw: 0 if raw == "Not_Reordered" else int(raw),
-    )
-    joining_type_stage1, joining_type_pages = generate_property(
-        joining_type_input,
-        data_parser=lambda raw: JOINING_TYPE_VALUES[raw],
-        missing_parser=lambda raw: JOINING_TYPE_FULL_NAMES[raw],
-    )
-    bidi_class_stage1, bidi_class_pages = generate_property(
-        bidi_class_input,
-        data_parser=lambda raw: BIDI_CLASS_VALUES[raw],
-        missing_parser=lambda raw: BIDI_CLASS_FULL_NAMES[raw],
-    )
-    script_stage1, script_pages = generate_property(
-        script_input,
-        data_parser=lambda raw: UNICODE_SCRIPT_VALUES[raw],
-        missing_parser=lambda raw: UNICODE_SCRIPT_VALUES[raw],
-    )
+    properties = [
+        ("COMBINING_CLASS", sys.argv[3],
+         build_value_map(aliases_path, "ccc")),
+        ("JOINING_TYPE", sys.argv[4],
+         build_value_map(aliases_path, "jt", JOINING_TYPE_ORDER)),
+        ("BIDI_CLASS", sys.argv[5],
+         build_value_map(aliases_path, "bc", BIDI_CLASS_ORDER)),
+        ("UNICODE_SCRIPT", sys.argv[6],
+         build_value_map(aliases_path, "sc", UNICODE_SCRIPT_ORDER,
+                         ignore=("Hrkt",))),
+    ]
 
     with open(output_path, "w") as output:
         output.write("#include <cstdint>\n\n")
         output.write("namespace {\n\n")
-        emit_property(
-            output, "COMBINING_CLASS", combining_class_stage1, combining_class_pages
-        )
-        emit_property(
-            output, "JOINING_TYPE", joining_type_stage1, joining_type_pages
-        )
-        emit_property(
-            output, "BIDI_CLASS", bidi_class_stage1, bidi_class_pages
-        )
-        emit_property(
-            output, "UNICODE_SCRIPT", script_stage1, script_pages
-        )
+        for prefix, input_path, value_map in properties:
+            stage1, pages = build_pages(parse_file(input_path, value_map))
+            emit_property(output, prefix, stage1, pages)
         output.write("} // namespace\n")
