+++ src/core/unicode/include/sourcemeta/core/unicode.h
@@ -5,6 +5,8 @@
 #include <sourcemeta/core/unicode_export.h>
 #endif
 
+#include <sourcemeta/core/unicode_ucd.h>
+
 #include <cstddef>     // std::size_t
 #include <cstdint>     // std::uint8_t
 #include <istream>     // std::istream
@@ -233,18 +235,6 @@ SOURCEMETA_CORE_UNICODE_EXPORT
 auto combining_class(const char32_t codepoint) -> std::uint8_t;
 
 /// @ingroup unicode
-/// The joining type of a Unicode codepoint per UAX #44. See
-/// https://www.unicode.org/reports/tr44/ for the property's definition.
-enum class JoiningType : std::uint8_t {
-  NonJoining = 0,
-  Transparent = 1,
-  LeftJoining = 2,
-  RightJoining = 3,
-  DualJoining = 4,
-  JoinCausing = 5,
-};
-
-/// @ingroup unicode
 /// Return the joining type of a Unicode codepoint. See
 /// https://www.unicode.org/reports/tr44/ for the property's definition.
 /// For example:
@@ -264,35 +254,6 @@ SOURCEMETA_CORE_UNICODE_EXPORT
 auto joining_type(const char32_t codepoint) -> JoiningType;
 
 /// @ingroup unicode
-/// The bidirectional class of a Unicode codepoint per UAX #44. See
-/// https://www.unicode.org/reports/tr44/ for the property's definition.
-enum class BidiClass : std::uint8_t {
-  LeftToRight = 0,
-  RightToLeft = 1,
-  ArabicLetter = 2,
-  EuropeanNumber = 3,
-  EuropeanSeparator = 4,
-  EuropeanTerminator = 5,
-  ArabicNumber = 6,
-  CommonSeparator = 7,
-  NonspacingMark = 8,
-  BoundaryNeutral = 9,
-  ParagraphSeparator = 10,
-  SegmentSeparator = 11,
-  WhiteSpace = 12,
-  OtherNeutral = 13,
-  LeftToRightEmbedding = 14,
-  LeftToRightOverride = 15,
-  RightToLeftEmbedding = 16,
-  RightToLeftOverride = 17,
-  PopDirectionalFormat = 18,
-  LeftToRightIsolate = 19,
-  RightToLeftIsolate = 20,
-  FirstStrongIsolate = 21,
-  PopDirectionalIsolate = 22,
-};
-
-/// @ingroup unicode
 /// Return the bidirectional class of a Unicode codepoint. See
 /// https://www.unicode.org/reports/tr44/ for the property's definition.
 /// For example:
@@ -310,187 +271,6 @@ enum class BidiClass : std::uint8_t {
 /// ```
 SOURCEMETA_CORE_UNICODE_EXPORT
 auto bidi_class(const char32_t codepoint) -> BidiClass;
-
-/// @ingroup unicode
-/// The script of a Unicode codepoint per UAX #24. See
-/// https://www.unicode.org/reports/tr24/ for the property's definition.
-enum class UnicodeScript : std::uint8_t {
-  Adlam = 0,
-  Ahom = 1,
-  AnatolianHieroglyphs = 2,
-  Arabic = 3,
-  Armenian = 4,
-  Avestan = 5,
-  Balinese = 6,
-  Bamum = 7,
-  BassaVah = 8,
-  Batak = 9,
-  Bengali = 10,
-  BeriaErfe = 11,
-  Bhaiksuki = 12,
-  Bopomofo = 13,
-  Brahmi = 14,
-  Braille = 15,
-  Buginese = 16,
-  Buhid = 17,
-  CanadianAboriginal = 18,
-  Carian = 19,
-  CaucasianAlbanian = 20,
-  Chakma = 21,
-  Cham = 22,
-  Cherokee = 23,
-  Chorasmian = 24,
-  Common = 25,
-  Coptic = 26,
-  Cuneiform = 27,
-  Cypriot = 28,
-  CyproMinoan = 29,
-  Cyrillic = 30,
-  Deseret = 31,
-  Devanagari = 32,
-  DivesAkuru = 33,
-  Dogra = 34,
-  Duployan = 35,
-  EgyptianHieroglyphs = 36,
-  Elbasan = 37,
-  Elymaic = 38,
-  Ethiopic = 39,
-  Garay = 40,
-  Georgian = 41,
-  Glagolitic = 42,
-  Gothic = 43,
-  Grantha = 44,
-  Greek = 45,
-  Gujarati = 46,
-  GunjalaGondi = 47,
-  Gurmukhi = 48,
-  GurungKhema = 49,
-  Han = 50,
-  Hangul = 51,
-  HanifiRohingya = 52,
-  Hanunoo = 53,
-  Hatran = 54,
-  Hebrew = 55,
-  Hiragana = 56,
-  ImperialAramaic = 57,
-  Inherited = 58,
-  InscriptionalPahlavi = 59,
-  InscriptionalParthian = 60,
-  Javanese = 61,
-  Kaithi = 62,
-  Kannada = 63,
-  Katakana = 64,
-  Kawi = 65,
-  KayahLi = 66,
-  Kharoshthi = 67,
-  KhitanSmallScript = 68,
-  Khmer = 69,
-  Khojki = 70,
-  Khudawadi = 71,
-  KiratRai = 72,
-  Lao = 73,
-  Latin = 74,
-  Lepcha = 75,
-  Limbu = 76,
-  LinearA = 77,
-  LinearB = 78,
-  Lisu = 79,
-  Lycian = 80,
-  Lydian = 81,
-  Mahajani = 82,
-  Makasar = 83,
-  Malayalam = 84,
-  Mandaic = 85,
-  Manichaean = 86,
-  Marchen = 87,
-  MasaramGondi = 88,
-  Medefaidrin = 89,
-  MeeteiMayek = 90,
-  MendeKikakui = 91,
-  MeroiticCursive = 92,
-  MeroiticHieroglyphs = 93,
-  Miao = 94,
-  Modi = 95,
-  Mongolian = 96,
-  Mro = 97,
-  Multani = 98,
-  Myanmar = 99,
-  Nabataean = 100,
-  NagMundari = 101,
-  Nandinagari = 102,
-  NewTaiLue = 103,
-  Newa = 104,
-  Nko = 105,
-  Nushu = 106,
-  NyiakengPuachueHmong = 107,
-  Ogham = 108,
-  OlChiki = 109,
-  OlOnal = 110,
-  OldHungarian = 111,
-  OldItalic = 112,
-  OldNorthArabian = 113,
-  OldPermic = 114,
-  OldPersian = 115,
-  OldSogdian = 116,
-  OldSouthArabian = 117,
-  OldTurkic = 118,
-  OldUyghur = 119,
-  Oriya = 120,
-  Osage = 121,
-  Osmanya = 122,
-  PahawhHmong = 123,
-  Palmyrene = 124,
-  PauCinHau = 125,
-  PhagsPa = 126,
-  Phoenician = 127,
-  PsalterPahlavi = 128,
-  Rejang = 129,
-  Runic = 130,
-  Samaritan = 131,
-  Saurashtra = 132,
-  Sharada = 133,
-  Shavian = 134,
-  Siddham = 135,
-  Sidetic = 136,
-  SignWriting = 137,
-  Sinhala = 138,
-  Sogdian = 139,
-  SoraSompeng = 140,
-  Soyombo = 141,
-  Sundanese = 142,
-  Sunuwar = 143,
-  SylotiNagri = 144,
-  Syriac = 145,
-  Tagalog = 146,
-  Tagbanwa = 147,
-  TaiLe = 148,
-  TaiTham = 149,
-  TaiViet = 150,
-  TaiYo = 151,
-  Takri = 152,
-  Tamil = 153,
-  Tangsa = 154,
-  Tangut = 155,
-  Telugu = 156,
-  Thaana = 157,
-  Thai = 158,
-  Tibetan = 159,
-  Tifinagh = 160,
-  Tirhuta = 161,
-  Todhri = 162,
-  TolongSiki = 163,
-  Toto = 164,
-  TuluTigalari = 165,
-  Ugaritic = 166,
-  Unknown = 167,
-  Vai = 168,
-  Vithkuqi = 169,
-  Wancho = 170,
-  WarangCiti = 171,
-  Yezidi = 172,
-  Yi = 173,
-  ZanabazarSquare = 174,
-};
 
 /// @ingroup unicode
 /// Return the script of a Unicode codepoint. See
